Rename steps and explicitly configure loading of them
